test: fix tnt_deauth() test case
The test case was mistakely changed in [1], because tarantool's
behaviour was incorrect for a while: it does not permit to authenticate
as 'guest' without a password. The tarantool's behaviour was fixed in
[2] and now the incorrect test case fails. The time to fix the test case
back.

Removed tarantool 1.7 and 1.9 from CI, because of two reasons: they are
not more supported, they show previous incorrect behaviour.

[1]: 0882011dac802b892517f6c4e26716c639da0d49 ('Fix a typo "is" ->
     "isnt" since previous connected test expect success')
[2]: https://github.com/tarantool/tarantool/issues/4327