cd: enable tarantool repo for rpmbuild/debuild
There are distributions, where tarantool is in base repositories (such
as Fedora 30-34 or Ubuntu Focal). There are ones, where it is in
semi-official repos such as EPEL (CentOS 7/8). There are ones, where we
have no tarantool package at all (or have only old tarantool-lts as in
Ubuntu Bionic).

It seems meaningful to enable the tarantool repository everywhere and
don't depend on presence of a tarantool package in distribution base
repositories.

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>