ci: actualize fast_testing workflow
This patch bumps the versions of Tarantool up to 2.11, 3.2 and 3.3 as
currently supported with built packages for the latest Ubuntu 24.04
(noble).

Also, it bumps the version of the `actions/checkout` actions to v4 and
the version of the `actions/setup-python` action to v5. It is needed for
fixing the following GitHub warnings:
| Node.js 16 actions are deprecated.
| Please update the following actions to use Node.js 20
|
| The following actions uses node12 which is deprecated and will be
| forced to run on node16

Also, it updates `setup-tarantool` action to v3. It is needed to make
Tarantool 3.x pipelines work, see [1] for details.

[1]: https://github.com/tarantool/setup-tarantool/releases/tag/v3.0.0